requirements.txt
adal==1.2.2
arrow==0.13.1
asn1crypto==1.0.1
astroid==2.0.1
azure-common==1.1.23
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.0
azure-mgmt-datafactory==0.8.0
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.1
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
binaryornot==0.4.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cookiecutter==1.6.0
cryptography==2.7
future==0.17.1
idna==2.8
isodate==0.6.0
isort==4.3.4
jinja2-time==0.2.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.6.10
msrestazure==0.6.2
oauthlib==3.1.0
poyo==0.4.2
pycparser==2.19
PyJWT==1.7.1
pylint==2.0.0
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.0
PyYAML==3.12
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.1.0
urllib3==1.24.2
whichcraft==0.5.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11